requirements.txt
kubernetes>=28.1.0
openshift-client>=1.0.18
pyyaml>=6.0
rich>=13.7.0
jinja2>=3.1.0
click>=8.1.0
